fix: resolve remaining 40+ compilation errors across 20 files

+++ app/src/main/java/com/aichathub/app/data/RemoteCatalogRepository.kt
@@ -201,7 +201,7 @@ class RemoteCatalogRepository(private val context: Context) {
             }
 
             val manifest = fetchManifest()
-            manifest.version > localVersion
+            manifest.version.toLong() > localVersion
         } catch (e: Exception) {
             Log.e(TAG, "Failed to check for updates", e)
             false
